pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Caussz", email = "cczcaua@gmail.com"},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
    "setuptools>=68.2.2",
    "requests>=2.31.0",
    "drf-spectacular>=0.27.1",
    "djangorestframework-simplejwt>=5.3.1",
    "matplotlib>=3.8.3",
    "django-cors-headers>=4.3.1",
    "django-extensions>=3.2.3",
    "pygraphviz>=1.12",
    "dotenv>=0.0.5",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.2",
    "psycopg2-binary>=2.9.9",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
    "Pillow>=10.4.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "djangorestframework>=3.14.0",
]

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"

